feat: reorganise functions serving logic
