+++ src/lib/string.js
@@ -1,0 +1,5 @@
+const capitalize = function (t) {
+  return t.replace(/(^\w|\s\w)/g, (string) => string.toUpperCase())
+}
+
+module.exports = { capitalize }
